refactor: moving files to separated directories
