+++ rs-react-app/src/components/ErrorBoundary.tsx
