+++ rs-react-app/src/style/ErrorBtn.css
